minor fixes player.cpp rework push function (not effective further modifications needed)

+++ Project_V3_SIAM_PASCAL_GERONDEAU/src/BoardGame.cpp
@@ -4,7 +4,7 @@
 //----------------------------------------CTOR-&-DTOR----------------------------------------//
 
 BoardGame::BoardGame()
-    :m_map(std::vector < std::vector<Piece*> >(5,std::vector<Piece*>(5,NULL)))
+    :m_map(std::vector < std::vector<Piece*> >(5,std::vector<Piece*>((unsigned int)5,NULL)))
 {
     for(int i=1;i<4;i++)
         m_map[i][2]= new Mountain(NULL);
